Rename pkg/parse to pkg/upstreamyaml
