+++ pkg/upstreamyaml/upstreamyaml.go
@@ -1,4 +1,4 @@
-package parse
+package upstreamyaml
 
 import (
 	"fmt"
